# openapi/spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/projects": {
      "post": {
        "description": "Create a new project with the specified name, description, and type",
        "operationId": "ProjectsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProjectResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Validation failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 422
                    },
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unprocessable Entity"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to create project",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Failed to create project"
                    },
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Create a new project",
        "tags": [
          "projects"
        ]
      },
      "get": {
        "description": "Get all projects for the organization with pagination",
        "operationId": "ProjectsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (default: 10)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projects retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProjectsResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 422
                    },
                    "message": {
                      "type": "string",
                      "example": "Organization not found"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unprocessable Entity"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to list projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Failed to list projects"
                    },
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "List all projects",
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{id}": {
      "get": {
        "description": "Retrieve a project by its unique identifier including its datasets",
        "operationId": "ProjectsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the project",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProjectResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Project not found"
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch project",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Failed to fetch project"
                    },
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get project by ID",
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{projectId}/upload": {
      "post": {
        "operationId": "ProjectsController_uploadFile",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadFileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadFileResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{id}/resources": {
      "post": {
        "description": "Create a new resource for the specified project",
        "operationId": "ProjectsController_createResource",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the project",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResourceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resource created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateResourceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Validation failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Project not found"
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 422
                    },
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unprocessable Entity"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to create resource",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Failed to create resource"
                    },
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Create resource for project",
        "tags": [
          "projects"
        ]
      },
      "get": {
        "description": "Get all resources for a project with signed download URLs and pagination",
        "operationId": "ProjectsController_listResources",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the project",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (default: 10)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resources retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResourcesResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Project not found"
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to list resources",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Failed to list resources"
                    },
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "List project resources",
        "tags": [
          "projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "SUPAEVAL API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the project",
            "example": "E-commerce Platform"
          },
          "description": {
            "type": "string",
            "description": "The description of the project",
            "example": "A modern e-commerce platform with advanced features and analytics"
          },
          "type": {
            "type": "string",
            "enum": [
              "MULTIMODAL",
              "IMAGE_EXTRACTION",
              "LLM_TEXT",
              "INSTANCE_SEGMENTATION"
            ],
            "description": "The type of the project",
            "example": "IMAGE_EXTRACTION"
          }
        },
        "required": [
          "name",
          "description",
          "type"
        ]
      },
...
      "CreateProjectResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the project",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "The name of the project",
            "example": "E-commerce Platform"
          },
          "description": {
            "type": "string",
            "description": "The description of the project",
            "example": "A modern e-commerce platform with advanced features and analytics"
          },
          "type": {
            "type": "string",
            "enum": [
              "MULTIMODAL",
              "IMAGE_EXTRACTION",
              "LLM_TEXT",
              "INSTANCE_SEGMENTATION"
            ],
            "description": "The type of the project",
            "example": "IMAGE_EXTRACTION"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "The creation timestamp",
            "example": "2024-03-14T12:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "The last update timestamp",
            "example": "2024-03-14T12:00:00.000Z"
          },
          "organizationId": {
            "type": "string",
            "description": "The organization ID that owns this project",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "createdById": {
            "type": "string",
            "description": "The user ID who created this project",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "datasets": {
            "description": "The datasets associated with this project",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DatasetEntity"
            }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "type",
          "createdAt",
          "updatedAt",
          "organizationId",
          "createdById"
        ]
      },
...
      "ListProjectsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Array of projects",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectEntity"
            }
          },
          "count": {
            "type": "number",
            "description": "Total count of items",
            "example": 25
          },
          "page": {
            "type": "number",
            "description": "Current page number",
            "example": 1
          },
          "nextPage": {
            "type": "number",
            "nullable": true,
            "description": "Next page number, null if no next page",
            "example": 2
          }
        },
        "required": [
          "items",
          "count",
          "page",
          "nextPage"
        ]
      },
...
      "UploadFileDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "description": "The name of the file to upload",
            "example": "image.jpg"
          },
          "contentType": {
            "type": "string",
            "description": "The MIME type of the file",
            "example": "image/jpeg"
          }
        },
        "required": [
          "fileName",
          "contentType"
        ]
      },
...
    }
  }
}